Doc revisions to the upgrade process
For issue: https://github.com/solo-io/gloo/issues/5393.

Includes doc style edits and reorganizing some of the info since it seemed a bit scattered.

Adds Enterprise example and shortcodes so that we can update the various GEE and GEOSS versions in one place and have that populate across the docs.