fix(tls_inspector): added AggregateListener support
"envoy.filters.listener.tls_inspector" is an essential listener_filter when attempting to use TLS.  For _most_ gateways, we would previously always consider logic on whether or not to add it based on the presence of a user-provided SslConfig field.  For the new _AggregateListener_, we neglected to add support, originally.  This caused TLS to be unable to function when using the "isolateVirtualHostsBySslConfig" feature flag.

related https://github.com/solo-io/gloo/issues/6677