Cherry-pick c511b2f6 onto master branch
Initial commit of tracing plugin changes

At this point, the plugin should in theory have all translation implemented. However, there is no error checking present in the plugin. A basic unit test exists for this work, but all it does it ensure that no errors occurred. The unit test also does not yet validate the content of the output object.