Final version of opencensus changes
This commit includes a translation fix and simplification of the API. Originally, we were outputting an object of type GrpcService_EnvoyGrpc_ for the grpcAddress option, but the Envoy plugin only supports the GrpcService_GoogleGrpc_ type. So we fix that here. We also simplify the protobuf API to use a oneof for specifying HTTP or GRPC.

I'm committing this now because we have decided to drop using OpenCensus due to the fact that it cannot be reconfigured. A lot of the work on this effort will be discarded, so I am saving it here for posterity.