Rewrite everything to use opentelemetry instead
So far, I have only tested manually, but it seems to work okay. One odd behaviour is that Envoy only seems to report information to the echo server on shutdown.

Unit tests aren't doing anything serious at the moment and need to be improved. It will also be necessary to write an e2e test for this. But otherwise, things seem to be working much more smoothly than they did with opencensus.