Remove race conditions from e2e tests
The major issue here was a pair of variables with near-global scope: collectorApiHit and collectionURLPath. Both of these were triggering go's race condition flags, probably because they were being written to and read from ginkgo's setup and teardown callbacks.

There was also an issue in shutting down the collection server, which is addressed by a comment around where the change was made.