Initial implementation with successful manual test
We create an environment variable in the gateway-proxy deployment and pass it to the `docker-entrypoint.sh` script which is the entrypoint for the deployment to start the envoy process. If this value is set to true, we call `ulimit -c 0` which disables core dumps. If it is set to false, then it is not called and core dumps can be created at will.

I have verified that this works as follows:

1. temporarily change the location core dumps are written to: `echo '/dev/shm/core.%e.%p' | sudo tee /proc/sys/kernel/core_pattern` - this probably only works on linux and hopefully it doesn't break your operating system, but i can't make any guarantees here

2. start a new cluster

3. `kubectl -n gloo-system edit deployments/gateway-proxy`

4. add the DISABLE_CORE_DUMPS environment variable and set it to "false" (quotes required)

5. `docker exec -it deployments/gateway-proxy sh`

6. send a signal to envoy signalling it to abort and leave a core dump `kill -SIGABRT $(pidof envoy)`

7. if you got kicked out of the container, run `docker exec -it deployments/gateway-proxy sh` again

8. `ls /dev/shm/` - should see the core dump file there

9. repeat steps 3-8 with DISABLE_CORE_DUMPS set to "true" and this time, no core files should appear in step 8

note that you can run this to check the status of whether core dumps are enabled: `cat /proc/$(pidof envoy)/environ | tr '\0' '\n' | grep CORE`

also note that in order to enable this, i had to put some debug code into `docker-entrypoint.sh` which i will remove in my next commit