Provide a default value for DISABLE_CORE_DUMPS
This isn't technically necessary, but helpful just in case something goes
wrong. In the previous iteration, the script would fail and envoy would not
start if DISABLE_CORE_DUMPS was not set. Technically, this should never happen
because the helm templates always provide this value. But if the entrypoint is
reached some other way (ie. not from the helm templates provided, or if the
environment variables in the deployment are removed for whatever reason), then
this script would fail and envoy would never start. So it seems safer to just
provide a default here.