Remove tap from global settings
This is a temporary change for the alpha release because we don't want to
decide right now exactly what behaviour we should use when settings get
overridden from a higher level (ie. settings) to a lower, more specific level
(ie. gateway).

We will revert this prior to the beta/final release once we have decided what
kind of behaviour we want to use for once and for all