add missing messages to status reports
This commit sets the 'message' field on various HTTPRoute conditions
to (hopefully) enable easier troubleshooting.

The translator case tests are also refactored to allow status/reporting
to be asserted along with (hopefully) making it easier and faster to
iterate on adding new cases.

Two new case tests are added to exercise the new messages added to
the HTTPRoute status messages.

Note: Forward port of (#9858)