.github: Update PR template to use comment formatting for instructions
This commit updates the pull request template to improve the usability
and clarity of the template. The changes involve converting the instructions
and guidelines under each section into comment formatting (`<!-- ... -->`).

This allows contributors to fill out the template without needing to delete
instructional text, reducing friction and ensuring that the template remains
clear and organized.

Signed-off-by: timflannagan <timflannagan@gmail.com>