[gateway2] Updates Gateway Controller Watch Predicate
Previously, the controller would only watch Gateway objects
for `generation` field changes which is not updated when annotations
change. Since Gateway reconciliation should be triggered when the
`gateway.gloo.solo.io/gateway-parameters-name` annotation is added,
removed, or modified, the predicate was updated to check for changes
in either the generation field or the annotations.

Signed-off-by: Daneyon Hansen <daneyon.hansen@solo.io>