Remove the test/consulvaulte2e e2e suite
The test/consulvaulte2e suite was one of the test suites that
cloudbuilder was configured to run on PRs.

See https://github.com/k8sgateway/k8sgateway/issues/10432 for
the issue that tracks whether we should migrate the other suite
being run (e.g. run-e2e-tests) and whether it's still relevant
for this project.

Signed-off-by: timflannagan <timflannagan@gmail.com>