install: Make initial bootstrap kgateway chart functional
The kgateway chart was initially bootstrapped via helm create
in a previous main branch PR. This chart is not functional and
contains some cruft that helm generates that isn't relevant to
this project (e.g. ingress). This commit deletes several files,
re-structures the values.yaml and templates/ directories.

Previous commits generated the required RBAC files. This expands
on that motion and introduces the GatewayClass and default
GatewayParameter CR templates.

Signed-off-by: timflannagan <timflannagan@gmail.com>