projects/gateway2/setup: Remove Settings CR from setup.yaml
This removes the Settings CR from the setup.yaml bootstrap
configuration in the envtests. The kgateway chart does not
carry this CR over to it's crds/ directory or have the RBAC
permissions to watch, list, etc. the resource.

Previously, we removed the legacy Settings logic in the gateway2
control plane in #10458, so this resource is no longer necessary
in the setup.yaml and envtest fails trying to deploy it with the
new chart.

Signed-off-by: timflannagan <timflannagan@gmail.com>