Remove the ci/github-actions directory
This commit removes the github-actions directory
from the repository. Previously, this directory
housed the go-test-summary and process-skip-directive
packages. With the want/need to shift away from the
current changelog format from the gloo project, the
latter is no longer relevant to this project.

I'm less familiar with the former and whether it
has any devex value that we should continue to
maintain with this project. I think we're in the
clear to remove it and re-introduce similiar functionality
when we hit pain points as the project continues to
mature.

Signed-off-by: timflannagan <timflannagan@gmail.com>