internal: Rename the gateway2 -> kgateway directory
The "gateway2" directory naming was an artifact from the
previous Gloo project where multiple gateway implementations
needed to be supported & is no longer relevant to this
project's scope.

When talking with a group of maintainers, there was less
of an appetite for a "controller" directory, in order to
remain consistent with the current helm chart naming,
underlying deployment/pod name, etc.

Signed-off-by: timflannagan <timflannagan@gmail.com>