e2e: Update BackendTLSPolicy certificates
The backend TLS policy suite is currently permafailing
due to an expired certificate. When reproducing locally
with a GatewayParameters resource that bumps the envoy
proxy logs to debug-level, we can see the following log
message:

```log
[2025-04-10 21:51:34.179][31][debug][connection] [external/envoy/source/common/tls/cert_validator/default_validator.cc:339] verify cert failed: X509_verify_cert: certificate verification error at depth 0: certificate has expired
```

Inspecting the certificate locally, it expired yesterday (04/09) and
needs to be re-generated. Since the private key is already committed,
we can re-generate a new certificate that expires a year from now.

```bash
$ # copy example.key in nginx.yaml to it's own file
$ openssl req -new -key test/kubernetes/e2e/features/backendtls/inputs/ca.key -out example.csr -subj "/CN=example.com"
$ openssl x509 -req -days 365 -in example.csr -signkey test/kubernetes/e2e/features/backendtls/inputs/ca.key -out example.crt | pbcopy
$ # manually update configmap.yaml & nginex.yaml certificates
```

This is basically a bandaid fix to unblock CI. In the future, we
may want to generate these keys in this suite's BeforeSuite logic.

Signed-off-by: timflannagan <timflannagan@gmail.com>