test: Remove the aws_credentials directory
A first pass towards cleaning up the test/ directory and any
legacy contents.

This commit removes the test/aws_credentials directory. Previously,
this suite contained tests that exercised the AWS services the gloo
project supported.

With the introduction of lambda support, and the adoption of localstack,
this suite is no longer relevant or consistent with the direction of this
project.

Related to https://github.com/kgateway-dev/kgateway/issues/11102.

Signed-off-by: timflannagan <timflannagan@gmail.com>