linter: forbid usage of kclient.New to honor discoveryNamespaceSelectors
Adds a linter rule to forbid the usage of kclient.New in favor
of kclient.NewFiltered to honor the discoveryNamespaceSelectors
ObjectFilter to filter out objects that are not relevant for
discovery. This will prevent accidental usage of kclient.New
on namespaced resources.

Although cluster scoped resources may use kclient.NewFiltered
with an ObjectFilter since the filter rule does not apply
to cluster scoped objects, it is acceptable to use kclient.New
with a //nolint:forbidigo directive.

Signed-off-by: Shashank Ram <shashank.ram@solo.io>