api/validation: use kubebuilder OneOf constraints
 Description
The OneOf constraints were inconsistent or incorrect.
This change updates controller-tools to use the newly
added ExactlyOneOf and AtMostOneOf validation rules
for objects with OneOf fields.

It also fixes the AwsBackend to require `lambda` since
that is the only supported backend and the code will
panic if `lambda` is unset.

 Change Type

```
/kind cleanup
```

 Changelog

```release-note
NONE
```

 Additional Notes
- Tested that CRDs can install with new CEL rules
  in k8s 1.28 and 1.32.
- Added a simple API validation test to validate
  API constraints.

Signed-off-by: Shashank Ram <shashank.ram@solo.io>