plugins/tp: Add PolicySubIR interface, refactor validation.go, etc
This commit continues the work established by the #11349 and #11791 refactor
PRs that aim to make the TrafficPolicy plugin more maintainable going forward.

Building on the sub-IR pattern introduced in #11791, this commits introduces
the PolicySubIR interface to formally standardize on this pattern and implement
a formal contract between the TrafficPolicyIR and each supported policy type
in the monolithic API.

Additionally, the validate.go file has been refactored, and each policy type
is now responsible for defining both an Equals and Validate method to satisfy
the new interface definition. For the latter, Validate() is purely scoped to
PGV-based validation. The validate.go file now contains a helper function that
runs the right validation pipeline based on the route replacement mode and
contains the xDS validation logic that is triggered with STRICT mode has been
enabled in the installation options.

As a result, a new Validate() method has been defined on the parent TrafficPolicy
type definition, which emulates the approach taken with the Equals() method were
it acts as a driver method and delegates to each sub-IR's Validate method.

Due to the lack of overall unit testing coverage (see #11785 for a formal tracker),
I introduce test cases for Equals() logic, and optionally, Validate() logic for
policy types that have restrictive PGV validation that's reproducible in unit
testing. I considered always adding test coverage that always asserts that
false, e.g. `assert.NoError(t, err, "Transformation validation should never fail at protobuf level")`.

Lastly, I considered expanding the initial interface with an explicit `Merge` or
`MergeInto` method, and refactor merge.go to move away from the helper function
approach in favor of the sub-IR method. However, this logic is new and potentially
less stable, etc.

Signed-off-by: timflannagan <timflannagan@gmail.com>