status: use Accepted=False,Reason=RouteRuleDropped for replaced rule
 Description
PartiallyInvalid condition has several problems:
1. It must be set only when PartiallyInvalid=True. This
   requires extra logic to invalidate an existing condition
   when there should be no PartiallyInvalid condition on
   an updated route status.
2. PartiallyInvalid must only be set when the route is partially
   invalid and not fully invalid. This requires post processing
   as translation merges routes with the same hostname and tracking
   whether a route is only partially invalid needs to group
   translated route rules by their parent route and check the rules
   for validity.

This change drops the usage of PartiallyInvalid=True in favor of
Accepted=False with Reason=RouteRuleDropped. A replaced route rule
will have this condition set. A route is replaced in case there are
processing errors or errors on the HTTPRouteRuleMatchIr, however the
condition is only set if there are processing errors from
runRoutePlugins. In the case of delegating/parent routes, it is expected
for the route to not have an Action set, in this case return early as
the rule simply delegates and is not relevant for the final output.

 Change Type

```
/kind cleanup
```

 Changelog

```release-note
NONE
```

 Additional Notes

Signed-off-by: Shashank Ram <shashank.ram@solo.io>