policy: remove unused AttachmentPoints construct
 Description
AttachmentPoints was initially introduced to control
which policies can attach to a targeted resource.
This is no longer required with CEL based enforcement
for targetRefs/targetSelectors, and is confusing to
use in various contexts. The only use case for this
is when there is a global policy the code wants to
explicitly opt-in to, but this isn't a scenario that
needs to be supported currently. Thus, this change
removes AttachmentPoints to simplify the code.

 Change Type

```
/kind cleanup
```

 Changelog

```release-note
NONE
```

 Additional Notes
Resolves #11775

Signed-off-by: Shashank Ram <shashank.ram@solo.io>