policy: deep-merging API for extAuth, extProc, transformations
 Description
Adds the ability to deep-merge extAuth, extProc, transformation
policies when multiple policies target the same resource.
Additionally, introduces the `kgateway.dev/policy-weight` annotation
for GatewayExtension and TrafficPolicy to control the priority
of policies during merging. All policies/extensions default to
a policy weight of 0, with a higher value resulting in greater
precedence.

Policies are merged in priority order and follow the following
precedence rules:
- By default, a policy with earlier creation timestamp is
  considered higher in priority and is preferred during a
  deep merge.
- For extAuth and extProc, the policy weight specified using
  the `kgateway.dev/policy-weight` annotation on the GatewayExtension
  determines the ordering of the extAuth and extProc filters relative
  to other policies during merging.
- For transformation, the policy weight specified using the
  `kgateway.dev/policy-weight` annotation on the TrafficPolicy
  determines the ordering of request/response transformation rules
  in the final rule slice.

 Change Type

```
/kind new_feature
```

 Changelog

```release-note
Enables optional deep merging of extAuth, extProc, transformation
policies in TrafficPolicy for policies attached to the same resource.
Enables the ability to prioritize policies and GatewayExtensions using
the kgateway.dev/policy-weight annotation.
```

 Additional Notes
Part of #11275

Signed-off-by: Shashank Ram <shashank.ram@solo.io>