translator: Add XListenerSet test case for invalid listener
We have an existing test case that exercises the two-way
handshake between XListenerSet and Gateway attachment, but
no explicit translator test case that tests a valid XLS
resource with an invalid listener. This edge case is only
covered in unit tests.

Adding this test case to primary test the new status logic
in the test/translator/test.go changes in the previous
commit.

Signed-off-by: timflannagan <timflannagan@gmail.com>