api/backend: fix CEL expression by escaping reserved namespace keyword
 Description
`namespace` is a reserved keyword in CEL and is not correctly
handled in k8s <= 1.31. Using `self.namespace` fails with
`undefined field 'namespace'`.

This change works around this bug by escaping the reserved
`namespace` keyword  using `__namespace__`. This is only
required in k8s <= 1.31, but since the CRDs are the same
for all versions, we use the escaping for always because
it works.

Tested that the Backend CRD can be installed without the
error seen previously.

 Change Type

```
/kind bug_fix
```

 Changelog

```release-note
FIX CEL rule evaluation in Backend API for k8s <= 1.31.
```

 Additional Notes
Resolves #12180

Signed-off-by: Shashank Ram <shashank.ram@solo.io>