translator: Ensure XLS HTTPS listeners are reported correctly
This updates the RC translator to ensure that invalid policy
attached to HTTPS-based listeners defined on an XLS resource
is reported correctly. Previously, the reporter would set
the status on the top-level Gateway conditions for this edge
case, which was a bug in the reportRouteConfigPolicyErrors
implementation.

The reportRouteConfigPolicyErrors function has been updated
to ensure XLS status reporting works now. Note: we cannot
re-use the policy.PolicyRef variable like we do to determine
whether the invalid policy attaches directly to listeners, but
the h.listener.PolicyAncestorRef which does have access to this
information at runtime.

Signed-off-by: timflannagan <timflannagan@gmail.com>