translator: Implement CDS validation
Follow-up to the #11939 PR that introduces RDS support to the
route config translation. This commit updates the backend
translator to additionally ensure the computed Envoy cluster
passes xDS validation when the controller is running in
strict mode.

This implies that envoy mode validate is being run twice during
a translation pass, and then per-IR event update for the
BackendConfigPolicy and TrafficPolicy plugins.

Implement xDS validation for the BackendConfigPolicy plugin to
act as a safety net between valid user configuration but invalid
xDS configuration that degrades the dataplane.

Note: While BackendTLSPolicy can contribute policy to the backend
being computed and that policy can lead to NACKs, then current
implementation needs more work and doesn't work with newer
versions of that experimental API or the stable version being
promoted in GW API 1.4.

Signed-off-by: timflannagan <timflannagan@gmail.com>