*: Add server-side TLS encryption for xDS
This commit introduces server-side TLS encryption for the xDS gRPC
server. When enabled, a user will provide a Secret of type `kubernetes.io/tls`
containing the server cert, key, and a `ca.crt` field via Helm values. The control
plane will mount and use this Secret when serving Envoy and agentgateway clients.

Additionally, the deployer component propagates the CA bundle to the proxy
so it can be used to validate the control plane's identity.

In order to support certificate rotation, the following changes were made:

- Kubelet updates the Secret mount symlinks on disk when the Secret data
  is rotated. The control plane runs an fsnotify watcher on the Secret-mounted
  certificate directory.
- When kubelet rotates the Secret (via symlink swap), the watcher reloads the
  certificate into the xDS gRPC server through the tls.Config GetCertificate
  callback. This helps avoid restarting the gRPC server
- At the same time, the watcher triggers reconciliation via a "kick channel" in
  the Gateway controller. This ensures the deployer re-reads the current
  ca.crt from disk and re-propagates the updated CA bundle to proxies
- Proxy pods additionally rely on kubelet’s Secret mount updates and for Envoy,
  the proxy is configured to use Envoy’s built-in [key rotation mechanism](https://www.envoyproxy.io/docs/envoy/v1.35.3/configuration/security/secret.html#key-rotation) to dynamically pick up the new CA bundle

This change is a follow-up to #12471, which introduced client-side JWT
authentication for xDS.

Signed-off-by: timflannagan <timflannagan@gmail.com>