flake: fix agentgateway status flake
Fixes https://github.com/kgateway-dev/kgateway/issues/12713

This is a recent regression. The root cause of the issue is a race:

* We enqueue a status write with no address
* The deployer writes the status
* We overwrite the status with no address

The fix here is to block the 3rd write from happening. Instead, we make
sure the object is unchanged since our status write (and re-enqueue it
if not).

Istio already takes core of re-enqueing it from a similar bug:
https://github.com/istio/istio/commit/b7742b9c37594c0c71bf10d652665d7e4b2dd1b8.

Signed-off-by: John Howard <john.howard@solo.io>