envoy: Add support for remaining gateway.spec.infrastructure
The envoy-based implementation currently supports the parametersRef
field for the Gateway infrastructure field, but doesn't currently
support the labels/annotations field under that same struct.

This is primarily a port of #12490 since that PR established the
API field documentation w.r.t. precedence and updated the primitives
in the deployer to make it easy to port over to the envoy-based class.

Infrastructure metadata is now propagated to all managed resources with
proper merging of GatewayParameters metadata. Infrastructure values take
precedence when keys conflict.

Signed-off-by: timflannagan <timflannagan@gmail.com>