hack: Add custom KRT IR equals linter
Introduce a custom golangci-lint linter that enforces that
KRT Equals() method implementations are correct. Leverages
the existing noKrtEquals marker that's documented but not
enforced to skip fields when checking underlying fields
for a struct that implements an Equals method.

Additionally, defines a special krtEqualsTODO marker for
existing code that will trigger linting violations.

Signed-off-by: timflannagan <timflannagan@gmail.com>