testutils: Enforce CRD structural schema
Update LoadFromFile to additionally enforce the CRD structural schema
for the input CR we're currently processing. Previously, we were
applying the CRD defaults to relevant objects, but if an input
YAML resource contained configuration that would lead to the apiserver
rejecting it in a real environment, this object was accepted and
the output golden file contained data that isn't representative of a real
environment.

Fixes #12832.

Signed-off-by: timflannagan <timflannagan@gmail.com>