Final removal of the deprecated envoy-based AI extension
This is a continuation of the #12690 and #12705 which removed
both the python-based ext_proc implementation and the underlying
Backend API plugin implementation.

Most of the testing has been unwinded at this point and there
was some initial changes towards updating the deployer to
no longer configure this component slated for removal on main.

This commit removes the remaining API fields and entirely removes
it from the deployer component.

Signed-off-by: timflannagan <timflannagan@gmail.com>